+++ crates/nostr-connect/src/client.rs
@@ -16,7 +16,6 @@ use tokio::sync::broadcast::Receiver;
 use tokio::sync::OnceCell;
 
 use crate::error::Error;
-use crate::util;
 
 /// Nostr Connect Client
 ///
@@ -259,7 +258,8 @@ impl NostrConnect {
             while let Ok(notification) = notifications.recv().await {
                 if let RelayPoolNotification::Event { event, .. } = notification {
                     if event.kind == Kind::NostrConnect {
-                        let msg: String = util::decrypt(secret_key, &event)?;
+                        let msg: String =
+                            nip44::decrypt(secret_key, &event.pubkey, event.content.as_str())?;
                         let msg: Message = Message::from_json(msg)?;
 
                         tracing::debug!("Received NIP46 message: '{msg}'");
@@ -414,7 +414,11 @@ async fn get_remote_signer_public_key(
             if let RelayPoolNotification::Event { event, .. } = notification {
                 if event.kind == Kind::NostrConnect {
                     // Decrypt content
-                    let msg: String = util::decrypt(app_keys.secret_key(), &event)?;
+                    let msg: String = nip44::decrypt(
+                        app_keys.secret_key(),
+                        &event.pubkey,
+                        event.content.as_str(),
+                    )?;
 
                     tracing::debug!("Received Nostr Connect message: '{msg}'");