+++ crates/nostr-connect/src/lib.rs
@@ -15,4 +15,3 @@ pub mod client;
 pub mod error;
 pub mod prelude;
 pub mod signer;
-mod util;